+++ client/src/components/Filter/Filter.jsx
@@ -1,0 +1,58 @@
+import React from 'react'
+import { useState } from "react"
+import "./Filter.css"
+
+export default function Filter() {
+
+    const [value, setValue] = useState()
+    const [budget, setBudget] = useState()
+    
+    const handleChange = (e) => {
+        const {id, value} = e.target;
+
+        setValue((prevInput) => ({
+            ...prevInput,
+            [id]: value,
+        }))
+        }
+    
+    const changeBudget = (e) => {
+        const {id, value} = e.target;
+
+        setBudget((prevInput) => ({
+            ...prevInput,
+            [id]: value,
+        }))
+    }
+
+    return (
+        <div className="filter-container">
+            <form>
+                <label>Filter By</label>
+                <select id="projectType" onChange={handleChange}>
+                    <option value="All">All</option>
+                    <option value="Bathroom">Bathroom</option>
+                    <option value="Ceiling">Ceiling</option>
+                    <option value="Electrical">Electrical</option>
+                    <option value="Floor">Floor</option>
+                    <option value="Kitchen">Kitchen</option>
+                    <option value="Landscape">Landscape</option>
+                    <option value="Paint">Paint</option>
+                    <option value="Plumbing">Plumbing</option>
+                    <option value="Roof">Roof</option>
+                    <option value="Wall">Wall</option>
+                </select>
+            </form>
+            <form>
+                <label>Filter By Budget</label>
+                <input 
+                    type="range"
+                    min="1"
+                    max="200000"
+                    value="100000"
+                    />
+            </form>
+
+        </div>
+    )
+}
